% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{ac}
\alias{ac}
\title{shortened as.character}
\usage{
ac()
}
\arguments{
\item{x}{object to be coerced or tested}

\item{...}{further arguments passed to or from other methods.}
}
\description{
shortened as.character
}
