% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{adds an alpha (transparency) to a color}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{A vector of colors}

\item{alpha}{The alpha level (0-1)}
}
\description{
adds an alpha (transparency) to a color
}
