% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{additivitetsPlot}
\alias{additivitetsPlot}
\title{Funktionen additivitetsPlot beregner gennemsnit i hver gruppe
defineret ved A:B (A og B er faktorer) og afsætter disse mod A.
For hvert niveau af B forbindes gennemsnit.
Desuden angives 0.95 CI for middelværdien}
\usage{
additivitetsPlot(A, B, x, ci = c("normal", "bootstrap", "none"))
}
\arguments{
\item{A}{faktor 1}

\item{B}{faktor 2}

\item{x}{response variable}

\item{ci}{type of ci (char)}
}
\value{
plots
}
\description{
Funktionen additivitetsPlot beregner gennemsnit i hver gruppe
defineret ved A:B (A og B er faktorer) og afsætter disse mod A.
For hvert niveau af B forbindes gennemsnit.
Desuden angives 0.95 CI for middelværdien
}
