% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{boolean_palette}
\alias{boolean_palette}
\title{boolean_palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
boolean_palette
}
\description{
Red/grey palette named vector for boolean variables
}
\keyword{datasets}
