% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{calculate_proportion}
\usage{
calculate_proportion(.tbl, var, group_var)
}
\arguments{
\item{.tbl}{a tibble}

\item{var}{the variable to calculate propotion of}

\item{group_var}{a grouping variable to calculate proportion within}
}
\value{
a tibble
}
\description{
calculate_proportion
}
