% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_bracket.R
\name{ci_bracket}
\alias{ci_bracket}
\title{Format confidence intervals}
\usage{
ci_bracket(x)
}
\arguments{
\item{no}{arguments necessary}
}
\description{
This function will combine columns from output of the confint function into a single formatted column
}
\examples{
lm <- lm(disp~hp, data = mtcars)
ci <- confint(lm)
ci_formatted <- ci_bracket(format_num(ci))
}
\keyword{confidence}
\keyword{interval}
