% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{constant_friendly_t_test}
\alias{constant_friendly_t_test}
\title{constant_friendly_t_test
Function returns a regular t.test except when vector is constant. When constant,
only "conf.int" element is returned}
\usage{
constant_friendly_t_test(vect)
}
\arguments{
\item{vect}{vector of values to perform t-test on}
}
\value{
double
}
\description{
constant_friendly_t_test
Function returns a regular t.test except when vector is constant. When constant,
only "conf.int" element is returned
}
