% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrees_minutes_seconds_convert.R
\name{dd_to_dms}
\alias{dd_to_dms}
\title{From decimal degrees to degrees minutes seconds}
\usage{
dd_to_dms(dd)
}
\arguments{
\item{dd}{vector of numbers  decimal degrees}
}
\value{
string vector of degrees minutes seconds
}
\description{
Convert degrees minutes seconds to decimal degrees
}
\examples{
myangle<- 45.15
dd_to_dms(myangle)
}
