% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_percentiles.R
\name{define_percentiles}
\alias{define_percentiles}
\title{For a df of FoodGroups and meal_portions in grams, split each food group 
into percentiles and assign percentile number to row
Highest percentile = TOP
Lowest percentile = BOTTOM}
\usage{
define_percentiles(df, percentiles = 5)
}
\arguments{
\item{df}{<- DF with columns 'FoodGroup"' and 'total_meal_portion'}

\item{percentiles}{<- Number of percentiles to split into, default=5}
}
\value{
x
}
\description{
For a df of FoodGroups and meal_portions in grams, split each food group 
into percentiles and assign percentile number to row
Highest percentile = TOP
Lowest percentile = BOTTOM
}
\author{
Robbie Pope
}
