% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{df_rem_outliers}
\alias{df_rem_outliers}
\title{df_rem_outliers}
\usage{
df_rem_outliers(df, col_name, percentage = 0.01)
}
\arguments{
\item{df}{data frame to remove outliers from}

\item{col_name}{which column to consider outliers for (char)}

\item{percentage}{which percentage to remove (in each tail) - Default = 0.01}
}
\value{
data frame without outliers
}
\description{
Shaves off rows where value of a given variable is an outlier
}
