% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrees_minutes_seconds_convert.R
\name{dms_to_dd}
\alias{dms_to_dd}
\title{From degrees minutes seconds to decimal degrees}
\usage{
dms_to_dd(s)
}
\arguments{
\item{s}{vector of characters in the form "degrees minutes seconds" separated by a blank space}
}
\value{
numeric vector in decimal degrees
}
\description{
Convert degrees minutes seconds to decimal degrees
}
\examples{
myangle<- "45  12    3.01"
dms_to_dd(myangle)
}
