% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bonferroni.R
\name{estimate.bonferroni}
\alias{estimate.bonferroni}
\title{Estimates the bonferroni corrected p-value threshold}
\usage{
estimate.bonferroni(meff, signific = 0.05)
}
\arguments{
\item{meff}{<- maximum effective number of tests (float)}

\item{signific}{<- Researcher set sginificance value, default is 0.05}
}
\value{
bonferroni corrected p value
}
\description{
Estimates the bonferroni corrected p-value threshold
}
\author{
Robbie Pope
}
