% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expected_shortfall}
\alias{expected_shortfall}
\title{expected_shortfall}
\usage{
expected_shortfall(v, q = 0.05)
}
\arguments{
\item{v}{vector (numeric)}

\item{q}{quantile for value-at-risk}
}
\value{
double
}
\description{
Calculate expected shortfall for a vector
}
