% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_mapview.R
\name{export_mapview}
\alias{export_mapview}
\title{export_mapview}
\usage{
export_mapview(map, string_filename = "file_name")
}
\arguments{
\item{map}{mapview object}

\item{string_filename}{string to name the output files}
}
\value{
mapview object
}
\description{
Export a mapview object to html and png files
}
