% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{Format p-values}
\usage{
format_p(x)
}
\arguments{
\item{suppress.p}{Select whether to suppress the "p" in the function output (FALSE by default)}
}
\description{
This function will print p-values according to APA-7 guidelines
}
\examples{
values <- c(0.2, 0.035, 0.00021)
format_p(values, suppress.p = TRUE)
}
\keyword{p-value}
