% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_condensed.R
\name{get_date_condensed}
\alias{get_date_condensed}
\title{get_date_condensed}
\usage{
get_date_condensed()
}
\value{
A string of today's date in the format yyyymmdd
}
\description{
Get the date as a codensed string in the format yyyymmdd,
without any white space.
}
