% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hush}
\alias{hush}
\title{hush}
\usage{
hush(code)
}
\arguments{
\item{code}{code chunk to hush}
}
\value{
Output of original function
}
\description{
Wrap another function call with hush() to silence output
}
