% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_transform.R
\name{inverse.transform.rank}
\alias{inverse.transform.rank}
\title{Inverse rank transform a vector}
\usage{
inverse.transform.rank(x)
}
\arguments{
\item{v}{<- vector of values}
}
\value{
col_inv <- a vector of values inverse transformed by rank
}
\description{
Inverse rank transform a vector
}
\author{
adapted from https://yingji15.github.io/2019-08-17-inverse-normal-transformation/
}
