% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Estimates the Jaccard similarity coefficient for two lists}
\usage{
jaccard_similarity(rslt1, rslt2)
}
\arguments{
\item{rslt1}{vector two  to compare}
}
\value{
: Jaccard similarity coefficient
}
\description{
Estimates the Jaccard similarity coefficient for two lists
}
\author{
Robbie Pope
}
