% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{linear_normal_model_test}
\alias{linear_normal_model_test}
\title{linear_normal_model_test}
\usage{
linear_normal_model_test(model, dat)
}
\arguments{
\item{model}{lm object}

\item{dat}{data frame from which model is generated}
}
\value{
plots
}
\description{
simple linear (normal univariate) model plots for model testing. Assuming dat only has columns
that are included in model, and every column in model is in dat
}
