% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lsmr.R
\name{lsmr}
\alias{lsmr}
\title{Solve the linear system Ax = b using lsmr}
\usage{
lsmr(A, b, lambda = 0, atol = 1e-05, btol = 1e-05, conlim = 1e+18,
  maxit = 50L, localSize = 0)
}
\arguments{
\item{A}{square matrix input}

\item{b}{numeric vector.}

\item{maxit}{integer. Maximum number of iterations}

\item{tol}{tolerance level}
}
\value{
list
}
\description{
Solve the linear system Ax = b using lsmr
}
\examples{
n <- 2000
p <- 10
lambda <- 0.1

set.seed(100)
x <- matrix(rnorm(n * p), ncol = p)
y <- x \%*\% rnorm(p) + rnorm(p)
b <- y
A <- tcrossprod(x) + lambda * diag(n)

system.time(alpha.true <- solve(A, b))
system.time(alpha.cg <- solveCG(A, b))

max(abs(alpha.cg$x - alpha.true))
}

