% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{make_coolors_url}
\alias{make_coolors_url}
\title{make_coolors_url}
\usage{
make_coolors_url(palette)
}
\arguments{
\item{palette}{A vector of colors}
}
\value{
A string that is a valid URL to paste into a browser
}
\description{
\code{make_coolors_url} function generates string
to use with the website coolors.co, which generates
aesthetically pleasing color palettes
}
