% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy_cols.R
\name{make_dummy_cols}
\alias{make_dummy_cols}
\title{make_dummy_cols}
\usage{
make_dummy_cols(data_df, id_var, dummy_var)
}
\arguments{
\item{data_df}{A data frame}

\item{id_var}{An identity variable to pivot against}

\item{dummy_var}{A dummy variable}
}
\value{
A data frame
}
\description{
Makes dummy columns in a data frame, given an ID variable
and a dummy variable
}
