% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_galaxy_plot}
\alias{make_galaxy_plot}
\title{make_galaxy_plot}
\usage{
make_galaxy_plot(
  .tbl,
  sample_n = 800,
  x = "UMAP_1",
  y = "UMAP_2",
  pointsize = 0,
  pixels = c(512, 512)
)
}
\arguments{
\item{.tbl}{a data frame or tibble with x and y coordinates}

\item{sample_n}{the number of points to display on top of the density map.
Set to 0 to remove all points}

\item{x}{string designating variable to plot on x-axis}

\item{y}{string designating variable to plot on y-axis}
}
\value{
ggplot object
}
\description{
Pretty sure Kevin Bi deserve the credit for this code but I've messed
around with it some too
}
