% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mark_significance}
\alias{mark_significance}
\title{mark_significance}
\usage{
mark_significance(x, level = 0.05)
}
\arguments{
\item{x}{char or double}

\item{level}{confidence level to mark - Default = 0.05}
}
\value{
char
}
\description{
Marks significance with an asterisk. Useful for both character and numeric
Always returns character
}
