% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boost_funcs.R
\name{multivarNorm}
\alias{multivarNorm}
\title{distribution function}
\usage{
multivarNorm(n = 1L, mean = 0, sigma, seed = NULL)
}
\arguments{
\item{n}{integer number of samples}

\item{mean}{vector mean of length equal to ncol(sigma) = nrow(sigma). If scalar, assumed the same for each dimension}

\item{sigma}{}

\item{seed}{integer; if NULL (default), seed is not forced.}
}
\value{
matrix of samples from multivariate normal distribution
}
\description{
distribution function
}
\examples{
library(rfunctions)
vals <- rnorm(100, mean = 4, sd = 2)
boost.pnorm(vals, mean = 4, sd = 2, lower.tail = FALSE)
}

