% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{neglog}
\alias{neglog}
\title{neglog}
\usage{
neglog(x, log_base = exp(1))
}
\arguments{
\item{x}{vector or integer}

\item{log_base}{base of log function - Default is exp(1)}
}
\value{
vector of transformed values
}
\description{
Log transforms a vector. Handles non-positive values. Injective and strictly increasing function
}
