% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{normal_control}
\alias{normal_control}
\title{normal_control}
\usage{
normal_control(vect)
}
\arguments{
\item{vect}{vector (numeric)}
}
\value{
None: Two plots
}
\description{
Simple normal qqplots and histogram for a vector
}
