% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{par_cor}
\alias{par_cor}
\title{par_cor}
\usage{
par_cor(dat, x, y, controls = c())
}
\arguments{
\item{dat}{data frame with variables}

\item{x}{column name of first variable (char)}

\item{y}{column name of second variable (char)}

\item{controls}{vector of column names of control variables}
}
\value{
double
}
\description{
calculated partial correlation of columns of data frame with specified controls
}
