% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea_bar.R
\name{plot_gsea_bar}
\alias{plot_gsea_bar}
\title{Plot top enriched GSEA pathways}
\usage{
plot_gsea_bar(fgsea_res, title, n_show = 20)
}
\arguments{
\item{fgsea_res}{Data.frame output of fgsea}

\item{title}{Plot title}

\item{n_show}{Number of pathways to display}
}
\value{
ggplot2 object
}
\description{
Plot top enriched GSEA pathways
}
