% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{plot_palette}
\alias{plot_palette}
\title{plot_palette}
\usage{
plot_palette(palette)
}
\arguments{
\item{palette}{A vector of colors}
}
\value{
A default R barplot
}
\description{
\code{plot_palette} function generates a barplot to visualize the palette colors
}
