% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Time function}
\usage{
plot_pca(
  pca_out,
  comp1,
  comp2,
  sep = NULL,
  col_index = NULL,
  str_extract_regex = NULL,
  col_label = ""
)
}
\arguments{
\item{pca_out}{output of prcomp}

\item{comp1}{integer value for PC on x-axis}

\item{comp2}{integer value for PC on y-axis}

\item{sep}{A string used for separating the sample name. Used with col_index}

\item{col_index}{An index specifying a particular piece of the separated sample name string. Used with sep}

\item{str_extract_regex}{A string specifying the regular expression to be extracted from the sample name for the color variable}

\item{col_label}{The label of the color variable}
}
\value{
A ggplot of a pca
}
\description{
plot_pca Plots a pca with a specific coloring scheme.
Also specifies the variance of each principal component.
}
