% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_proportion}
\alias{plot_proportion}
\title{plot_proportion}
\usage{
plot_proportion(.tbl, x = group, y = prop, fill = clusters)
}
\arguments{
\item{.tbl}{a tibble}

\item{x}{symbol designating column of tibble with group information}

\item{y}{symbol designating column of tibble with proportion values}

\item{fill}{symbol designating fill color}
}
\value{
ggplot object
}
\description{
plot_proportion
}
