% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_two_groups}
\alias{plot_two_groups}
\title{plot_two_groups}
\usage{
plot_two_groups(
  .tbl,
  group,
  var = prop * 100,
  stat.method = "wilcox.test",
  color_palette = c("black", "lightblue")
)
}
\arguments{
\item{.tbl}{a data frame or tibble}

\item{group}{group to plot as symbol

Use with fig.asp = 2, fig.width = 2.}

\item{var}{variable to plot as symbol}
}
\value{
ggplot object
}
\description{
plot_two_groups
}
