% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_violin}
\alias{plot_violin}
\title{plot_violin}
\usage{
plot_violin(.tbl, group, var, pt.size = 0, pt.stroke = 1)
}
\arguments{
\item{.tbl}{a data frame or tibble with x and y coordinates}

\item{group}{group to plot as symbol}

\item{var}{variable to plot as symbol}
}
\value{
ggplot object
}
\description{
plot_violin
}
