% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano
Generates a pretty volcano plot, by default takes in dataframes
formatted like the output of Seurat::FindMarkers
Pretty by my standards}
\usage{
plot_volcano(
  volcano_df,
  var_fc = avg_log2FC,
  var_pval = p_val_adj,
  var_gene = gene,
  thresh = Inf
)
}
\arguments{
\item{volcano_df}{data frame with volcano data}

\item{var_fc}{fold change variable}

\item{thresh}{pval-threshold of where to display gene names up to. Can be
a single number, which will set a single cutoff, or a numerical vector of
length 2, which will set two cutoffs on the left and the right side of the
volcano}

\item{var_p}{pvalue variable}
}
\value{
a ggplot2 object displaying a pretty volcano plot
}
\description{
plot_volcano
Generates a pretty volcano plot, by default takes in dataframes
formatted like the output of Seurat::FindMarkers
Pretty by my standards
}
