% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gene_rank.R
\name{prepare_gene_rank}
\alias{prepare_gene_rank}
\title{Prepare a named vector of gene ranks for GSEA}
\usage{
prepare_gene_rank(res)
}
\arguments{
\item{res}{DESeq2 result object with stat column}
}
\value{
Named numeric vector: names = genes, values = ranking statistic
}
\description{
Prepare a named vector of gene ranks for GSEA
}
