% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rem_outliers}
\alias{rem_outliers}
\title{rem_outliers}
\usage{
rem_outliers(vect, percentage = 0.01)
}
\arguments{
\item{vect}{vector}

\item{percentage}{which percentage to shave off each tail. Default = 0.01}
}
\value{
vector without outliers
}
\description{
Shaves a certain percentage off each tail (of vector)
}
