% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove.outliers}
\alias{remove.outliers}
\title{Set outliers to NA by two different methods
Standard deviations frmo the mean or by interquartile range}
\usage{
remove.outliers(v, method = c("sd", "iqr"), n.sd = 3)
}
\arguments{
\item{v}{vector of values to process}

\item{method}{sd is standard deviation and iqr is interquartile range}

\item{n.sd}{number of SD from mean, only relevant if method='sd'}
}
\value{
vector with outliers set to NA based on selected method
}
\description{
Set outliers to NA by two different methods
Standard deviations frmo the mean or by interquartile range
}
\author{
Adapted from Alessia Visconti
}
