% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deseq_results.R
\name{run_deseq_results}
\alias{run_deseq_results}
\title{Run DESeq2 differential expression with Wald test and ashr shrinkage}
\usage{
run_deseq_results(dds, contrast)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{contrast}{Character vector of length 3 specifying contrast (e.g. c("group", "SOS", "WT"))}
}
\value{
A DESeqResults object with LFC shrinkage and preserved Wald statistic
}
\description{
Run DESeq2 differential expression with Wald test and ashr shrinkage
}
