% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fgsea.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Run fgsea enrichment}
\usage{
run_fgsea(gene_rank, pathways, minSize = 10, maxSize = 500, nperm = 10000)
}
\arguments{
\item{gene_rank}{Named numeric vector}

\item{pathways}{List of gene sets}

\item{minSize}{Minimum gene set size}

\item{maxSize}{Maximum gene set size}

\item{nperm}{Number of permutations}
}
\value{
fgsea result data.frame
}
\description{
Run fgsea enrichment
}
