% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_or_readRDS.R
\name{run_or_readRDS}
\alias{run_or_readRDS}
\title{run_or_readRDS}
\usage{
run_or_readRDS(path, f, verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{path}{a path to look for an RDS object with the saved results}

\item{f}{a function that will output results that you want saved}

\item{verbose}{boolean specifying verbosity}
}
\value{
a tibble
}
\description{
A kind of hack cache-ing function that I use in my notebook for
convenience. Will read in an RDS object at the specified path if it exists,
otherwise will run the specified function.
}
