% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlot.R
\name{savePlot}
\alias{savePlot}
\title{Save a ggplot as PNG and PDF}
\usage{
savePlot(
  filename,
  plot,
  tag = tag,
  plots_dir = plots_dir,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{filename}{Base filename (no extension)}

\item{plot}{A ggplot object}

\item{tag}{Subfolder name}

\item{width}{Plot width (optional)}

\item{height}{Plot height (optional)}
}
\description{
Save a ggplot as PNG and PDF
}
