% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{social_palette}
\alias{social_palette}
\title{social_palette}
\format{
An object of class \code{character} of length 29.
}
\usage{
social_palette
}
\description{
Palette with colors taken from social media logos
}
\keyword{datasets}
