% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer_lme4.R
\name{stargazer_lme4}
\alias{stargazer_lme4}
\title{stargazer function for lme4 with random effects}
\usage{
stargazer_lme4(..., type = "latex", title = "", style = "default",
  summary = NULL, out = NULL, out.header = FALSE, column.labels = NULL,
  column.separate = NULL, covariate.labels = NULL, dep.var.caption = NULL,
  dep.var.labels = NULL, dep.var.labels.include = TRUE, align = FALSE,
  coef = NULL, se = NULL, t = NULL, p = NULL, t.auto = TRUE,
  p.auto = TRUE, ci = FALSE, ci.custom = NULL, ci.level = 0.95,
  ci.separator = NULL, add.lines = NULL, apply.coef = NULL,
  apply.se = NULL, apply.t = NULL, apply.p = NULL, apply.ci = NULL,
  colnames = NULL, column.sep.width = "5pt", decimal.mark = NULL,
  df = TRUE, digit.separate = NULL, digit.separator = ",", digits = 2,
  digits.extra = NULL, flip = FALSE, float = TRUE, float.env = "table",
  font.size = NULL, header = TRUE, initial.zero = NULL,
  intercept.bottom = TRUE, intercept.top = FALSE, keep = NULL,
  keep.stat = NULL, label = "", model.names = NULL,
  model.numbers = TRUE, multicolumn = TRUE, no.space = NULL,
  notes = NULL, notes.align = NULL, notes.append = TRUE,
  notes.label = NULL, object.names = FALSE, omit = NULL,
  omit.labels = NULL, omit.stat = NULL, omit.summary.stat = NULL,
  omit.table.layout = NULL, omit.yes.no = c("Yes", "No"), order = NULL,
  ord.intercepts = FALSE, perl = FALSE, report = NULL, rownames = NULL,
  rq.se = "nid", selection.equation = FALSE, single.row = FALSE,
  star.char = NULL, star.cutoffs = NULL, suppress.errors = FALSE,
  table.layout = NULL, table.placement = "!htbp", zero.component = FALSE,
  summary.logical = TRUE, summary.stat = NULL, nobs = TRUE,
  mean.sd = TRUE, min.max = TRUE, median = FALSE, iqr = FALSE)
}
\description{
stargazer function for lme4 with random effects
}

