\name{subtract-methods}
\docType{methods}
\alias{subtract-methods}
\alias{subtract,ANY,ANY-method}
\alias{subtract,dgCMatrix,dgCMatrix-method}
\alias{subtract,dgCMatrix,dsCMatrix-method}
\alias{subtract,dsCMatrix,dgCMatrix-method}
\alias{subtract,dsCMatrix,dsCMatrix-method}
\alias{subtract,matrix,matrix-method}
\title{Methods for Function \code{subtract}}
\description{
Methods for function \code{subtract}. \cr
subtract(a, b) == b - a

}
\section{Methods}{
\describe{

\item{\code{signature(a = "ANY", b = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(a = "dgCMatrix", b = "dgCMatrix")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(a = "dgCMatrix", b = "dsCMatrix")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(a = "dsCMatrix", b = "dgCMatrix")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(a = "dsCMatrix", b = "dsCMatrix")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(a = "matrix", b = "matrix")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
