% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_subtract.R
\name{subtract}
\alias{subtract}
\title{subtract b from a}
\usage{
subtract(a, b)
}
\arguments{
\item{a}{matrix input}

\item{b}{matrix input}
}
\value{
A matrix equal to b - a
}
\description{
subtract b from a
}

