% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sum_outliers}
\alias{sum_outliers}
\title{sum_outliers}
\usage{
sum_outliers(dat, outlier_col, sum_col, percentage = 0.1)
}
\arguments{
\item{dat}{Data frame}

\item{outlier_col}{column name (char) to define which observations are considered outliers}

\item{sum_col}{which column (char) to sum ouliers for}

\item{percentage}{which percentage of each tail is considered outliers}
}
\value{
double
}
\description{
Sums values of specified column of outliers and non-outliers of another specified column
}
