% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_function.R
\name{time_function}
\alias{time_function}
\title{Time function}
\usage{
time_function(my_func, ...)
}
\arguments{
\item{my_func}{A function}
}
\value{
A time difference
}
\description{
\code{time_function} function uses \code{Sys.time()} to calculate how long a
function takes to run
}
