% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_in_spreadsheet.R
\name{view_in_spreadsheet}
\alias{view_in_spreadsheet}
\title{Helper function to view a dataset with default spreadsheet}
\usage{
view_in_spreadsheet(.data)
}
\arguments{
\item{.data}{a dataset}
}
\value{
nothing
}
\description{
View a dataset with default spreadsheet, taken and slightly adapted from:
https://twitter.com/brodriguesco/status/1447468259725434886?s=20
}
\examples{
mtcars |>
view_in_spreadsheet()

library(magrittr)
mtcars \%>\%
view_in_spreadsheet()

}
