% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_convert.R
\name{wdws2uv}
\alias{wdws2uv}
\title{convert wind direction and wind speed to u and v components}
\usage{
wdws2uv(wd, ws)
}
\arguments{
\item{wd}{a vector of wind directions}

\item{ws}{a vector of wind speeds}
}
\value{
a matrix
}
\description{
this function converts wind direction and wind speed to u and v components
}
\examples{
set.seed(123)
wd <- as.integer(rnorm(10, 180, 90))

set.seed(123)
ws <- rnorm(10, 4, 1)

## convert to u and v
wdws2uv(wd, ws)

## convert back
uv <- wdws2uv(wd, ws)
uv2wdws(uv[, 1], uv[, 2])

}
\seealso{
\code{\link{uv2wdws}}
}
\author{
Tim Appelhans
https://raw.githubusercontent.com/environmentalinformatics-marburg/Rsenal/master/R/wdws2uv.R
}
