% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/products.R
\name{crossprodChunk}
\alias{crossprodChunk}
\title{Compute X'X in chunks}
\usage{
crossprodChunk(X, row.chunk = NULL, sparse = FALSE)
}
\arguments{
\item{X}{matrix input}

\item{row.chunk}{integer. Number of rows per chunk. Specify NULL for no chunking}

\item{sparse}{Boolean value to specify if matrix X is sparse}
}
\value{
A symmetric matrix X'X
}
\description{
Compute X'X in chunks
}
\examples{
crossprodChunk(matrix(rnorm(1e4 * 1e2), 1e4, 1e2), row.chunk = 500, FALSE)
}

