% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_to_dd.R
\name{dms_to_dd}
\alias{dms_to_dd}
\title{dms_to_dd}
\usage{
dms_to_dd(s)
}
\arguments{
\item{s}{a vector of strings in the form "degrees minutes seconds" separated by a blank space}
}
\value{
a numeric vector in decimal degrees
}
\description{
Transform degrees minutes seconds to decimal degrees
}
\examples{
myangle<- "45  12    3.01"
dms_to_dd(myangle)
}
