% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{Format p-values}
\usage{
format_p(x)
}
\arguments{
\item{no}{arguments necessary}
}
\description{
This function will print p-values according to APA-7 guidelines
}
\examples{
values <- c(0.2, 0.035, 0.00021)
vapply(values, format_p, character(1))
}
\keyword{p-value}
