% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lsmr.R
\name{lsmr}
\alias{lsmr}
\title{Solve the linear system A'Ax = Ab using lsmr}
\usage{
lsmr(A, b, lambda = 0, atol = 1e-05, btol = 1e-05, conlim = 1e+18,
  maxit = 50L, localSize = 0)
}
\arguments{
\item{A}{square matrix input}

\item{b}{numeric vector.}

\item{maxit}{integer. Maximum number of iterations}

\item{tol}{tolerance level}
}
\value{
list
}
\description{
Solve the linear system A'Ax = Ab using lsmr
}
\examples{
n <- 10000
p <- 100

set.seed(100)
x <- simSparseMatrix(0.8, boolean = FALSE, dim = c(n, p))
b <- as.vector(x \%*\% rnorm(p) + rnorm(n))

A <- crossprod(x)
xtb <- as.vector(crossprod(x, b))

system.time( alpha.true <- solve(A, b) )
system.time( alpha.lsmr <- lsmr(x, b) )

max(abs(alpha.lsmr$x - alpha.true))
}

