% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_group_split.R
\name{named_group_split}
\alias{named_group_split}
\title{named_group_split
credit goes to romainfrancois on https://github.com/tidyverse/dplyr/issues/4223}
\usage{
named_group_split(.tbl, ..., sep_str = " / ")
}
\arguments{
\item{.tbl}{a tibble}
}
\value{
a tibble
}
\description{
named_group_split
credit goes to romainfrancois on https://github.com/tidyverse/dplyr/issues/4223
}
